refactor: move tag routes to new module
