+++ retrorecon/routes/__init__.py
@@ -11,5 +11,6 @@ from .dagdotdev import bp as dagdotdev_bp
 from .urls import bp as urls_bp
 from .swagger import bp as swagger_bp
 from .overview import bp as overview_bp
+from .tags import bp as tags_bp
 
-__all__ = ['notes_bp', 'tools_bp', 'db_bp', 'settings_bp', 'domains_bp', 'docker_bp', 'registry_bp', 'dag_bp', 'oci_bp', 'dagdotdev_bp', 'urls_bp', 'swagger_bp', 'overview_bp']
+__all__ = ['notes_bp', 'tools_bp', 'db_bp', 'settings_bp', 'domains_bp', 'docker_bp', 'registry_bp', 'dag_bp', 'oci_bp', 'dagdotdev_bp', 'urls_bp', 'swagger_bp', 'overview_bp', 'tags_bp']
